requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
yt-dlp>=2025.1.1
PyYAML>=6.0.2
pydantic>=2.8.0
